gateway/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.2.0
bcrypt==4.3.0
certifi==2025.1.31
click==8.1.8
comm==0.2.2
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
dnspython==2.7.0
dotenv==0.9.9
ecdsa==0.19.1
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.11
frozenlist==1.5.0
gotrue==2.11.4
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
matplotlib-inline==0.1.7
multidict==6.1.0
nest-asyncio==1.6.0
packaging==24.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
platformdirs==4.3.6
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pyzmq==26.2.1
realtime==2.4.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.1
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
